Prepare to impl init cmd and path mod commands
